Support WriteTextFile and CopyFile as built-in actions
Replace the WriteTextFile machinery with a more general mechanism for 'built-in actions' - if a node doesn't specify an action type then by default it will be a RunShellCommand action, but otherwise it can specify a typecode for one of the built-in actions. WriteTextFile itself has been converted to a built-in action. 

CopyFile has also been added as a built-in action (and implemented for OSX and Win32). It must have exactly one input and one output.